+++ tests/LocalDateTest.php
@@ -132,6 +132,12 @@ class LocalDateTest extends AbstractTestCase
         $this->assertLocalDateIs(2018, 7, 21, LocalDate::fromDateTime($dateTime));
     }
 
+    public function testFromNativeDateTime(): void
+    {
+        $dateTime = new \DateTime('2018-07-21');
+        $this->assertLocalDateIs(2018, 7, 21, LocalDate::fromNativeDateTime($dateTime));
+    }
+
     /**
      * @dataProvider providerNow
      *
@@ -1403,12 +1409,42 @@ class LocalDateTest extends AbstractTestCase
      * @param string $dateTime The date-time string that will be parse()d by LocalDate.
      * @param string $expected The expected output from the native DateTime object.
      */
+    public function testToNativeDateTime(string $dateTime, string $expected): void
+    {
+        $zonedDateTime = LocalDate::parse($dateTime);
+        $dateTime = $zonedDateTime->toNativeDateTime();
+
+        $this->assertInstanceOf(\DateTime::class, $dateTime);
+        $this->assertSame($expected, $dateTime->format('Y-m-d\TH:i:s.uO'));
+    }
+
+    /**
+     * @dataProvider providerToDateTime
+     *
+     * @param string $dateTime The date-time string that will be parse()d by LocalDate.
+     * @param string $expected The expected output from the native DateTime object.
+     */
     public function testToDateTime(string $dateTime, string $expected): void
     {
         $zonedDateTime = LocalDate::parse($dateTime);
         $dateTime = $zonedDateTime->toDateTime();
 
         $this->assertInstanceOf(\DateTime::class, $dateTime);
+        $this->assertSame($expected, $dateTime->format('Y-m-d\TH:i:s.uO'));
+    }
+
+    /**
+     * @dataProvider providerToDateTime
+     *
+     * @param string $dateTime The date-time string that will be parse()d by LocalDate.
+     * @param string $expected The expected output from the native DateTime object.
+     */
+    public function testToNativeDateTimeImmutable(string $dateTime, string $expected): void
+    {
+        $zonedDateTime = LocalDate::parse($dateTime);
+        $dateTime = $zonedDateTime->toNativeDateTimeImmutable();
+
+        $this->assertInstanceOf(\DateTimeImmutable::class, $dateTime);
         $this->assertSame($expected, $dateTime->format('Y-m-d\TH:i:s.uO'));
     }