+++ CHANGELOG.md
@@ -22,6 +22,9 @@ The following methods have been deprecated in favour of new names ([#47](https:/
 - `LocalTime::toDateTime()` → `toNativeDateTime()`
 - `LocalTime::toDateTimeImmutable()` → `toNativeDateTimeImmutable()`
 - `LocalDateRange::toDatePeriod()` → `toNativeDatePeriod()`
+- `LocalDate::fromDateTime()` → `fromNativeDateTime()`
+- `LocalDate::toDateTime()` → `toNativeDateTime()`
+- `LocalDate::toDateTimeImmutable()` → `toNativeDateTimeImmutable()`
 
 ## [0.4.0](https://github.com/brick/date-time/releases/tag/0.4.0) - 2021-12-23